Allow 'runsc do' to run without root
'--rootless' flag let's a non-root user execute 'runsc do'.
The drawback is that the sandbox and gofer processes will
run as root inside a network namespace that is mapped to the
caller's user, intead of nobody. And only network is defaulted
to '--network=host' inside the root network namespace. On the
bright side, it's very convenient for testing:

runsc --rootless do ls
runsc --rootless do curl www.google.com

PiperOrigin-RevId: 252150357